Gate jobs were updated
git-rebase step was added to all Gate jobs.
Now proposed patchset is rebased to current master before running main
job's script.

+++ scripts/git-rebase.sh
@@ -1,5 +1,13 @@
 #!/bin/bash -e
 
+cd $WORKSPACE
 git fetch origin master
-git rebase master
-git clean -x -f -d
+
+if [ $? != 0 ]
+then
+    echo "ERROR. git fetch failed"
+    return 1
+else
+   git rebase master
+   git clean -x -f -d
+fi
